Move code to brainglobe-utils
