--- brainglobe_workflows/brainmapper/tools/tools.py
@@ -1,49 +1,0 @@
-from natsort import natsorted
-
-
-def check_unique_list(in_list, natural_sort=True):
-    """
-    Checks if all the items in a list are unique or not
-    :param list in_list: Input list
-    :param bool natural_sort: Sort the resulting items naturally
-    (default: True)
-    :return: True/False and a list of any repeated values
-    """
-    unique = set(in_list)
-    repeated_items = []
-
-    for item in unique:
-        count = in_list.count(item)
-        if count > 1:
-            repeated_items.append(item)
-
-    if repeated_items:
-        if natural_sort:
-            repeated_items = natsorted(repeated_items)
-        return False, repeated_items
-    else:
-        return True, []
-
-
-def common_member(a, b, natural_sort=True):
-    """
-    Checks if two lists (or sets) have a common member, and if so, returns
-    the common members.
-    :param a: First list (or set)
-    :param b: Second list (or set)
-    :param bool natural_sort: Sort the resulting items naturally
-    (default: True)
-    :return: True/False and the list of values
-    """
-    a_set = set(a)
-    b_set = set(b)
-    intersection = list(a_set.intersection(b_set))
-    if len(intersection) > 0:
-        result = True
-    else:
-        result = False
-
-    if natural_sort:
-        intersection = natsorted(intersection)
-
-    return result, intersection
