+++ tests/brainmapper/test_unit/test_tools/test_prep.py
@@ -3,12 +3,7 @@ import os
 import pytest
 from brainglobe_utils.general.exceptions import CommandLineInputError
 
-from brainglobe_workflows.brainmapper.tools import prep
-
-# import shutil
-
-# from pathlib import Path
-
+from brainglobe_workflows.brainmapper import prep
 
 data_dir = os.path.join("tests", "data")
 
@@ -92,146 +87,6 @@ class Paths:
         self.tmp__cubes_output_dir = cubes_output_dir
         self.classification_out_file = classification_out_file
         self.cells_in_standard_space = cells_in_standard_space
-
-
-# def get_dict_of_what_to_run(what_to_run):
-#     what_to_run_dict = {
-#         "detect": what_to_run.detect,
-#         "classify": what_to_run.classify,
-#         "register": what_to_run.register,
-#         "figures": what_to_run.figures,
-#     }
-#     return what_to_run_dict
-#
-#
-# def test_calc_what_to_run(tmpdir):
-#     tmpdir = str(tmpdir)
-#
-#     # registered_atlas_path = os.path.join(tmpdir, "registered_atlas.tiff")
-#     cells_file_path = os.path.join(tmpdir, "cells.xml")
-#     classification_out_file_path = os.path.join(
-#         tmpdir, "cell_classification.xml"
-#     )
-#     summary_cell_counts_path = os.path.join(
-#     tmpdir, "summary_cell_counts.csv")
-#     figures_dir = os.path.join(tmpdir, "figures")
-#
-#     args = Args(
-#         output_dir=tmpdir,
-#         registration_output_folder=tmpdir,
-#         cells_file_path=cells_file_path,
-#         classification_out_file=classification_out_file_path,
-#         figures_dir=figures_dir,
-#     )
-#
-#     # default
-#     what_to_run = prep.CalcWhatToRun(args)
-#     default_validation = {
-#         "detect": True,
-#         "classify": True,
-#         "register": False,
-#         "figures": False,
-#     }
-#     default_test = get_dict_of_what_to_run(what_to_run)
-#     assert default_validation == default_test
-#
-#     # options
-#     args.no_classification = True
-#     what_to_run = prep.CalcWhatToRun(args)
-#     validation = {
-#         "detect": True,
-#         "classify": False,
-#         "register": False,
-#         "figures": False,
-#     }
-#     test = get_dict_of_what_to_run(what_to_run)
-#     assert validation == test
-#
-#
-#     # # existance
-#     # Path(registered_atlas_path).touch()
-#     # what_to_run = prep.CalcWhatToRun(args)
-#     # validation = {
-#     #     "detect": True,
-#     #     "classify": True,
-#     #     "register": False,
-#     #     "figures": True,
-#     # }
-#     # test = get_dict_of_what_to_run(what_to_run)
-#     # assert validation == test
-#
-#     Path(cells_file_path).touch()
-#     what_to_run.update(args)
-#     validation = {
-#         "detect": False,
-#         "classify": True,
-#         "register": False,
-#         "figures": True,
-#     }
-#     test = get_dict_of_what_to_run(what_to_run)
-#     assert validation == test
-#
-#     os.remove(cells_file_path)
-#     what_to_run.update(args)
-#     validation = {
-#         "detect": True,
-#         "classify": True,
-#         "register": False,
-#         "figures": True,
-#     }
-#     test = get_dict_of_what_to_run(what_to_run)
-#     assert validation == test
-#
-#     Path(classification_out_file_path).touch()
-#     what_to_run.update(args)
-#     validation = {
-#         "detect": False,
-#         "classify": False,
-#         "register": False,
-#         "figures": True,
-#     }
-#     test = get_dict_of_what_to_run(what_to_run)
-#     assert validation == test
-#
-#     Path(summary_cell_counts_path).touch()
-#     what_to_run.update(args)
-#     validation = {
-#         "detect": False,
-#         "classify": False,
-#         "register": False,
-#         "summarise": False,
-#         "figures": True,
-#     }
-#     test = get_dict_of_what_to_run(what_to_run)
-#     assert validation == test
-#
-#     make_and_fill_directory(figures_dir)
-#     args.figures = True
-#     what_to_run.update(args)
-#     validation = {
-#         "detect": False,
-#         "classify": False,
-#         "register": False,
-#         "summarise": False,
-#         "figures": False,
-#     }
-#     test = get_dict_of_what_to_run(what_to_run)
-#     assert validation == test
-#
-#
-#     # os.remove(registered_atlas_path)
-#     os.remove(summary_cell_counts_path)
-#     shutil.rmtree(figures_dir)
-#     what_to_run.update(args)
-#     validation = {
-#         "detect": False,
-#         "classify": False,
-#         "register": True,
-#         "summarise": True,
-#         "figures": True,
-#     }
-#     test = get_dict_of_what_to_run(what_to_run)
-#     assert validation == test
 
 
 def make_and_fill_directory(directory):
